requirements.txt
django
djangorestframework
mysqlclient
boto3
django-cors-headers
python-decouple
djangorestframework-simplejwt
psycopg2